+++ src/component/page/bruker-detaljer/BrukerDetaljerPage.tsx
@@ -2,7 +2,7 @@ import { AlertStripeFeil } from 'nav-frontend-alertstriper';
 import React from 'react';
 import { useParams } from 'react-router-dom';
 
-import { fetchTiltakDeltagerDetaljer } from '../../../api/tiltakAPI';
+import { fetchTiltakDeltagerDetaljer } from '../../../api/tiltak-api';
 import { Spinner } from '../../felles/spinner/Spinner';
 import { BrukerPaaTiltakDetaljer } from './BrukerPaaTiltakDetaljer';
 import { isNotStartedOrPending, isRejected, usePromise } from '../../../utils/use-promise';
